+++ src/includes/class-ajax.php
@@ -2130,35 +2130,6 @@ class SUPER_Ajax {
                 if( $v['type']=='files' ) {
                     if( ( isset( $v['files'] ) ) && ( count( $v['files'] )!=0 ) ) {
                         foreach( $v['files'] as $key => $value ) {
-                            
-                            // // Check that the nonce is valid, and the user can edit this post.
-                            // if ( 
-                            //     isset( $_POST['my_image_upload_nonce'], $_POST['post_id'] ) 
-                            //     && wp_verify_nonce( $_POST['my_image_upload_nonce'], 'my_image_upload' )
-                            //     && current_user_can( 'edit_post', $_POST['post_id'] )
-                            // ) {
-                            //     // The nonce was valid and the user has the capabilities, it is safe to continue.
-
-                            //     // These files need to be included as dependencies when on the front end.
-                            //     require_once( ABSPATH . 'wp-admin/includes/image.php' );
-                            //     require_once( ABSPATH . 'wp-admin/includes/file.php' );
-                            //     require_once( ABSPATH . 'wp-admin/includes/media.php' );
-                                
-                            //     // Let WordPress handle the upload.
-                            //     // Remember, 'my_image_upload' is the name of our file input in our form above.
-                            //     $attachment_id = media_handle_upload( 'my_image_upload', $_POST['post_id'] );
-                                
-                            //     if ( is_wp_error( $attachment_id ) ) {
-                            //         // There was an error uploading the image.
-                            //     } else {
-                            //         // The image was uploaded successfully!
-                            //     }
-
-                            // } else {
-
-                            //     // The security check failed, maybe show the user an error.
-                            // }
-
                             // Before we proceed check if the file already exists, if so, do nothing
                             // Exclude files that are being uploaded for the first time
                             // They will be in the "uploads/php" directory
@@ -2174,54 +2145,214 @@ class SUPER_Ajax {
                             $file = basename( $value['url'] );
                             $folder = basename( dirname( $value['url'] ) );
                             
-                            // @since 3.1 - skip if one of the values are empty
+                            // Skip if one of the values are empty
                             if( ($file=='') || ($folder=='') ) continue;
 
-                            $path = SUPER_PLUGIN_DIR . '/uploads/php/files/' . $folder . '/' . $file;
-                            
-                            // @since 1.3
-                            // Make sure to skip this file if it's source location is invalid
-                            if (strpos($path, 'uploads/php/files') !== false) {
+                            // Get source file
+                            $sourcePath = SUPER_PLUGIN_DIR . '/uploads/php/files/' . $folder . '/' . $file;
+                            $sourcePath = urldecode( $sourcePath );
 
-                                $source = urldecode( $path );
-                                $wp_upload_dir = wp_upload_dir();
-                                $folder = $wp_upload_dir['basedir'] . '/superforms' . $wp_upload_dir["subdir"];
-                                $unique_folder = SUPER_Common::generate_random_folder($folder);
-                                $newfile = $unique_folder . '/' . basename( $source );
-                                if ( !copy( $source, $newfile ) ) {
-                                    $dir = str_replace( basename( $source ), '', $source );
-                                    SUPER_Common::delete_dir( $dir );
-                                    SUPER_Common::delete_dir( $unique_folder );
-                                    SUPER_Common::output_message(
-                                        $error = true,
-                                        $msg = esc_html__( 'Failed to copy', 'super-forms' ) . '"'.$source.'" to: "'.$newfile.'"',
-                                        $redirect = $redirect
-                                    );
-                                    die();
-                                }else{
-                                    $dir = str_replace( basename( $source ), '', $source );
-                                    if( !empty( $dir ) ) {
-                                        $delete_dirs[] = $dir;
-                                    }
-                                    $filename = $newfile;
-                                    $filetype = wp_check_filetype( basename( $filename ), null );
-                                    $wp_upload_dir = wp_upload_dir();
-                                    $attachment = array(
-                                        'post_mime_type' => $filetype['type'],
-                                        'post_title'     => preg_replace( '/\.[^.]+$/', '', basename( $filename ) ),
-                                        'post_content'   => '',
-                                        'post_status'    => 'inherit'
-                                    );
-                                    $attach_id = wp_insert_attachment( $attachment, $filename );
+                            // Determine location to store the file
+                            $wp_upload_dir = wp_upload_dir();
 
-                                    require_once( ABSPATH . 'wp-admin/includes/image.php' );
-                                    $attach_data = wp_generate_attachment_metadata( $attach_id, $filename );
-                                    wp_update_attachment_metadata( $attach_id,  $attach_data );
-                                    
-                                    $data[$k]['files'][$key]['attachment'] = $attach_id;
-                                    $data[$k]['files'][$key]['url'] = wp_get_attachment_url( $attach_id );
-                                }
+                            // Store in custom directory, or in a secure folder
+                            // The user can define this in `Super Forms > Settings > File Upload Settings`
+
+                            // Default directory (/wp-content/uploads/super-forms):
+                            $folder = $wp_upload_dir['basedir'] . '/superforms' . $wp_upload_dir["subdir"];
+                            $unique_folder = SUPER_Common::generate_random_folder($folder);
+                            $newfile = $unique_folder . '/' . basename( $sourcePath );
+                            $dir = str_replace( basename( $sourcePath ), '', $sourcePath );
+                            if ( !copy( $sourcePath, $newfile ) ) {
+                                $errors = error_get_last();
+                                SUPER_Common::delete_dir( $dir );
+                                SUPER_Common::delete_dir( $unique_folder );
+                                SUPER_Common::output_message(
+                                    $error = true,
+                                    $msg = '<strong>' . esc_html__( 'Upload failed', 'super-forms' ) . ':</strong> ' . $errors['message'],
+                                    $redirect = $redirect
+                                );
+                                die();
+                            }else{
+                                if( !empty( $dir ) ) $delete_dirs[] = $dir;
+                                $filetype = wp_check_filetype( basename( $newfile ), null );
+                                $attachment = array(
+                                    'post_mime_type' => $filetype['type'],
+                                    'post_title'     => preg_replace( '/\.[^.]+$/', '', basename( $newfile ) ),
+                                    'post_content'   => '',
+                                    'post_status'    => 'inherit'
+                                );
+                                $attach_id = wp_insert_attachment( $attachment, $newfile );
+                                require_once( ABSPATH . 'wp-admin/includes/image.php' );
+                                $attach_data = wp_generate_attachment_metadata( $attach_id, $newfile );
+                                wp_update_attachment_metadata( $attach_id,  $attach_data );
+                                $data[$k]['files'][$key]['attachment'] = $attach_id;
+                                $data[$k]['files'][$key]['url'] = wp_get_attachment_url( $attach_id );
+                                
+                                // Add to Media library?
+                                // $addToMediaLibrary = false;
+                                // if($addToMediaLibrary){
+                                //     // Do not make the file visilbe in Media Library
+                                //     // This can be done by not attaching it to a post
+                                //     $url = site_url();
+                                //     $url = rtrim($url, '/') . '/';
+                                //     $url = str_replace(wp_normalize_path( ABSPATH ), $url, wp_normalize_path( $newfile ));
+                                //     $url = esc_url_raw( $url );
+                                //     $data[$k]['files'][$key]['attachment'] = 0;
+                                //     $data[$k]['files'][$key]['url'] = $url;
+                                // }else{
+
+                                //}
+                                ///$data[$k]['files'][$key]['attachment'] = 0;
+                                //$data[$k]['files'][$key]['url'] = wp_get_attachment_url( $attach_id );
+                                //$attach_id = wp_insert_attachment( $attachment, $newfile );
+                                //error_log('$attach_id: ' . $attach_id, 0);
+                                // require_once( ABSPATH . 'wp-admin/includes/image.php' );
+                                // $attach_data = wp_generate_attachment_metadata( $attach_id, $newfile );
+                                // error_log('$attach_data: ' . json_encode($attach_data), 0);
+                                // wp_update_attachment_metadata( $attach_id,  $attach_data );
+                                //$data[$k]['files'][$key]['attachment'] = $attach_id;
+                                //$data[$k]['files'][$key]['url'] = wp_get_attachment_url( $attach_id );
                             }
+
+                            // if( !$entry_id ) return;
+                            // //wp_delete_attachment( $attachment_id, true );
+                            // $attachments = get_posts( array(
+                            //     'post_type' => 'attachment',
+                            //     'posts_per_page' => -1,
+                            //     'post_parent' => $entry_id
+                            // ) );
+                            // if ( $attachments ) {
+                            //     foreach ( $attachments as $attachment ) {
+                            //         $uploads = wp_get_upload_dir();
+                            //         $file_meta = wp_get_attachment_metadata( $attachment->ID );
+                            //         if( $file_meta && isset($file_meta['sizes']) ){
+                            //             foreach ($file_meta['sizes'] as $key => $meta) {
+                            //                 @unlink( trailingslashit( $uploads['path'] ) . $meta['file'] );
+                            //             }
+                            //         }
+                            //         $delete = get_attached_file( $attachment->ID );
+                            //         @unlink( $delete );
+                            //         wp_delete_attachment( $attachment->ID, true );
+                            //     }
+                            // }
+
+                            // error_log('$newfile: ' . $newfile, 0);
+                            // // Secure file directory
+                            // $filedir = apply_filters('super_forms_secure_dir_filter', SUPER_PLUGIN_SECURE_DIR);
+                            // error_log('$filedir: ' . $filedir, 0);
+                            // $newfile = $filedir . basename( $sourcePath );
+                            // error_log('$newfile2: ' . $newfile, 0);
+                            // // Create directory if not yet exists
+                            // if (!file_exists($filedir)) {
+                            //     error_log('filedir1', 0);
+                            //     mkdir($filedir, 0744, true);
+                            // }
+                            // if (!file_exists($filedir)) {
+                            //     error_log('filedir2', 0);
+                            //     mkdir($filedir, 0744, true);
+                            // }else{
+                            //     error_log('exists!', 0);
+                            // }
+
+                            // if ( !copy( $sourcePath, $newfile ) ) {
+                            //     $dir = str_replace( basename( $sourcePath ), '', $sourcePath );
+                            //     SUPER_Common::delete_dir( $dir );
+                            //     SUPER_Common::delete_dir( $unique_folder );
+                            //     SUPER_Common::output_message(
+                            //         $error = true,
+                            //         $msg = esc_html__( 'Failed to copy', 'super-forms' ) . '"'.$sourcePath.'" to: "'.$newfile.'"',
+                            //         $redirect = $redirect
+                            //     );
+                            //     die();
+                            // }else{
+                            //     $dir = str_replace( basename( $sourcePath ), '', $sourcePath );
+                            //     if( !empty( $dir ) ) {
+                            //         $delete_dirs[] = $dir;
+                            //     }
+                            //     $filename = $newfile;
+                            //     $filetype = wp_check_filetype( basename( $filename ), null );
+                            //     $wp_upload_dir = wp_upload_dir();
+                            //     $attachment = array(
+                            //         'post_mime_type' => $filetype['type'],
+                            //         'post_title'     => preg_replace( '/\.[^.]+$/', '', basename( $filename ) ),
+                            //         'post_content'   => '',
+                            //         'post_status'    => 'inherit'
+                            //     );
+                            //     $attach_id = wp_insert_attachment( $attachment, $filename );
+
+                            //     require_once( ABSPATH . 'wp-admin/includes/image.php' );
+                            //     $attach_data = wp_generate_attachment_metadata( $attach_id, $filename );
+                            //     wp_update_attachment_metadata( $attach_id,  $attach_data );
+                                
+                            //     $data[$k]['files'][$key]['attachment'] = $attach_id;
+                            //     $data[$k]['files'][$key]['url'] = wp_get_attachment_url( $attach_id );
+                            // }
+                        
+
+                        // // Before we proceed check if the file already exists, if so, do nothing
+                        // // Exclude files that are being uploaded for the first time
+                        // // They will be in the "uploads/php" directory
+                        // $file = $value['url'];
+                        // if(!strpos($file, 'uploads/php/files')) {
+                        //     $file_headers = @get_headers($file);
+                        //     if($file_headers && $file_headers[0] != '404') {
+                        //         continue;
+                        //     }
+                        // }
+
+                        // // If the file does not exists let's process it
+                        // $file = basename( $value['url'] );
+                        // $folder = basename( dirname( $value['url'] ) );
+                        
+                        // // @since 3.1 - skip if one of the values are empty
+                        // if( ($file=='') || ($folder=='') ) continue;
+
+                        // $path = SUPER_PLUGIN_DIR . '/uploads/php/files/' . $folder . '/' . $file;
+                        
+                        // // @since 1.3
+                        // // Make sure to skip this file if it's source location is invalid
+                        // if (strpos($path, 'uploads/php/files') !== false) {
+
+                        //     $sourcePath = urldecode( $path );
+                        //     $wp_upload_dir = wp_upload_dir();
+                        //     $folder = $wp_upload_dir['basedir'] . '/superforms' . $wp_upload_dir["subdir"];
+                        //     $unique_folder = SUPER_Common::generate_random_folder($folder);
+                        //     $newfile = $unique_folder . '/' . basename( $sourcePath );
+                        //     if ( !copy( $sourcePath, $newfile ) ) {
+                        //         $dir = str_replace( basename( $sourcePath ), '', $sourcePath );
+                        //         SUPER_Common::delete_dir( $dir );
+                        //         SUPER_Common::delete_dir( $unique_folder );
+                        //         SUPER_Common::output_message(
+                        //             $error = true,
+                        //             $msg = esc_html__( 'Failed to copy', 'super-forms' ) . '"'.$sourcePath.'" to: "'.$newfile.'"',
+                        //             $redirect = $redirect
+                        //         );
+                        //         die();
+                        //     }else{
+                        //         $dir = str_replace( basename( $sourcePath ), '', $sourcePath );
+                        //         if( !empty( $dir ) ) {
+                        //             $delete_dirs[] = $dir;
+                        //         }
+                        //         $filename = $newfile;
+                        //         $filetype = wp_check_filetype( basename( $filename ), null );
+                        //         $wp_upload_dir = wp_upload_dir();
+                        //         $attachment = array(
+                        //             'post_mime_type' => $filetype['type'],
+                        //             'post_title'     => preg_replace( '/\.[^.]+$/', '', basename( $filename ) ),
+                        //             'post_content'   => '',
+                        //             'post_status'    => 'inherit'
+                        //         );
+                        //         $attach_id = wp_insert_attachment( $attachment, $filename );
+
+                        //         require_once( ABSPATH . 'wp-admin/includes/image.php' );
+                        //         $attach_data = wp_generate_attachment_metadata( $attach_id, $filename );
+                        //         wp_update_attachment_metadata( $attach_id,  $attach_data );
+                                
+                        //         $data[$k]['files'][$key]['attachment'] = $attach_id;
+                        //         $data[$k]['files'][$key]['url'] = wp_get_attachment_url( $attach_id );
+                        //     }
+                        // }
                         }
                     }
                 }else{