+++ src/includes/class-common.php
@@ -1792,8 +1792,10 @@ class SUPER_Common {
         if( $global_settings['smtp_enabled']=='disabled' ) {
             $wpmail_attachments = array();
             foreach( $attachments as $k => $v ) {
+                $v = str_replace('https://', 'http://', $v );
                 $path = str_replace(str_replace('https://', 'http://', content_url()), '', $v);
-                $wpmail_attachments[] = WP_CONTENT_DIR . $path;
+                $filesystemPath = WP_CONTENT_DIR . $path;
+                $wpmail_attachments[] = $filesystemPath;
             }
 
             SUPER_Forms()->session->set( 'super_string_attachments', $string_attachments );